Refactor test support builders
Split broad shared test-support helpers into focused builder modules while keeping the existing import facades stable.
