+++ tools/dev/oversized_test_allowlist.yml
@@ -8,8 +8,6 @@ allowlist:
   apps/server/tests/adapters/pdf/test_report_document_projection.py: High-density report projection regression matrix kept together by document-output seam.
   # Settings smoke keeps the full multi-tab settings journey in one browser spec.
   apps/ui/tests/smoke.settings.spec.ts: End-to-end settings smoke journey intentionally spans multiple settings tabs in one spec.
-  # Report helpers are the shared builder/assertion spine for multiple PDF and report test modules.
-  apps/server/tests/test_support/report_helpers.py: Shared report test helper surface reused across multiple report and PDF suites.
   # Metrics log helper coverage still lives as one dense helper-focused module.
   apps/server/tests/use_cases/run/test_metrics_log_helpers.py: Helper-focused metrics log regression coverage intentionally stays aggregated in one module.
   # Maintenance mounted-DOM coverage now lives in one Node-side regression owner instead of four legacy bridge specs.
